chore: memoized attached notes
